Moved code related to conversion from Composite definitions on GraphQL SDL to Internal Composite Definitions to a separate module
